+++ scripts/giant_file_registry.toml
@@ -570,10 +570,3 @@ file = "src/services/discord/catch_up.rs"
 owner = "discord-relay"
 deadline = "2026-10-31"
 decompose_issue = "#3405"
-
-[[entry]]
-# Crossed 1000-prod-LoC at #4159 r3 (offset-provenance DeferredClaim gating).
-file = "src/services/discord/tui_direct_abort_marker/mod.rs"
-owner = "discord-relay"
-deadline = "2026-10-31"
-decompose_issue = "#4175"
