+++ src/services/discord/tui_direct_abort_marker/mod.rs
@@ -212,10 +212,20 @@ impl AbortedAnchorMarker {
 // ---------------------------------------------------------------------------
 
 mod deferred_claim;
+mod drain;
 mod store;
+mod sweep;
+mod tombstone;
 
 pub(in crate::services::discord) use deferred_claim::{
     LiveInflightProbe, ensure_marker_for_own_synthetic_turn, record_for_deferred_claim,
+};
+#[cfg(test)]
+use drain::drain_on_terminal_commit_with_applier_and_offsets;
+pub(in crate::services::discord) use drain::{
+    drain_on_terminal_commit, drain_on_terminal_commit_with_applier,
+    drain_on_terminal_commit_with_applier_and_offset, drain_on_terminal_commit_with_offset,
+    drain_on_terminal_commit_with_offsets, resolve_own_claim_markers_for_visibly_completed_anchor,
 };
 use store::reload;
 #[cfg(test)]
@@ -231,6 +241,10 @@ pub(in crate::services::discord) use store::{
 };
 #[cfg(test)]
 use store::{root, tombstone_root};
+pub(in crate::services::discord) use sweep::{sweep_expired, sweep_expired_with_applier};
+pub(in crate::services::discord) use tombstone::{
+    cover_from_commit_tombstone, post_abort_commit_tombstone, record_for_abort,
+};
 
 fn is_zero_id_commit_for_real_deferred_claim_marker(
     marker: &AbortedAnchorMarker,
@@ -565,69 +579,6 @@ fn deferred_claim_live_inflight_is_pinned(
         )
 }
 
-/// Record-instant 대조 (codex r2 — replaces the unfounded "pre-covered"
-/// promotion): a tombstone satisfying the marker's commit-evidence gate
-/// ALREADY durable at record time means that turn terminal-committed before
-/// this marker existed (its drain pass could never see it) → stamp covered
-/// with the commit instant. The id-0-for-real-marker carve-out is the only
-/// exception: because id-0 is not an identity, that tombstone must be recorded
-/// at-or-after this marker was created.
-pub(super) fn cover_from_commit_tombstone(marker: &mut AbortedAnchorMarker) -> bool {
-    if marker.covered_at_ms.is_some() {
-        return false;
-    }
-    let Some(t) = load_commit_tombstones(&marker.provider, marker.channel_id)
-        .into_iter()
-        .find(|t| commit_tombstone_matches_marker(marker, t))
-    else {
-        return false;
-    };
-    marker.covered_at_ms = Some(t.committed_at_ms);
-    true
-}
-
-/// Sweep-side 대조: a tombstone covers an UNCOVERED marker when it satisfies
-/// the marker's commit-evidence gate AND its commit is not EARLIER than the
-/// abort. `>=`, not `>` (codex r3): identity/start evidence is PRIMARY — the
-/// ABORT fired because the recorded foreign turn stayed live through the whole
-/// 32s backstop, so a commit of THAT turn cannot predate the input submission
-/// (itself before the marker existed); an evidence-matched same-ms commit is
-/// therefore necessarily post-submission, and strict `>` only ever rejected
-/// ANSWERED anchors at the ms boundary (false `⚠`). The record-instant 대조
-/// above owns evidence predating the marker; this subsidiary guard mirrors the
-/// drain's.
-pub(super) fn post_abort_commit_tombstone(marker: &AbortedAnchorMarker) -> Option<u64> {
-    load_commit_tombstones(&marker.provider, marker.channel_id)
-        .into_iter()
-        .find(|t| {
-            commit_tombstone_matches_marker(marker, t) && t.committed_at_ms >= marker.aborted_at_ms
-        })
-        .map(|t| t.committed_at_ms)
-}
-
-/// The ABORT path's one-call record: build the marker (always uncovered —
-/// codex r2), run the record-instant tombstone 대조, persist. Returns the
-/// persisted marker for the caller's structured log fields.
-pub(super) fn record_for_abort(
-    provider: String,
-    channel_id: u64,
-    anchor_message_id: u64,
-    tmux_session_name: String,
-    foreign: Option<(u64, String)>,
-) -> Result<AbortedAnchorMarker, String> {
-    let mut marker = AbortedAnchorMarker::for_abort(
-        provider,
-        channel_id,
-        anchor_message_id,
-        tmux_session_name,
-        now_ms(),
-        foreign,
-    );
-    cover_from_commit_tombstone(&mut marker);
-    record(&marker)?;
-    Ok(marker)
-}
-
 // ---------------------------------------------------------------------------
 // Pure decision functions (truth-table tested — no I/O, no clock)
 // ---------------------------------------------------------------------------
@@ -883,423 +834,6 @@ pub(super) fn shared_reaction_applier(shared: Arc<SharedData>) -> ReactionApplie
 
 fn now_ms() -> u64 {
     chrono::Utc::now().timestamp_millis().max(0) as u64
-}
-
-/// Watcher terminal-commit chokepoint: a body-visible normal commit covers
-/// every marker whose recorded foreign identity matches the COMMITTED turn
-/// (codex r1) → `⏳ → ✅`. Returns markers drained.
-pub(super) async fn drain_on_terminal_commit(
-    shared: &Arc<SharedData>,
-    provider: &str,
-    tmux_session_name: &str,
-    channel_id: u64,
-    committed_user_msg_id: u64,
-    committed_started_at: &str,
-) -> usize {
-    drain_on_terminal_commit_with_offset(
-        shared,
-        provider,
-        tmux_session_name,
-        channel_id,
-        committed_user_msg_id,
-        committed_started_at,
-        None,
-    )
-    .await
-}
-
-pub(super) async fn drain_on_terminal_commit_with_offset(
-    shared: &Arc<SharedData>,
-    provider: &str,
-    tmux_session_name: &str,
-    channel_id: u64,
-    committed_user_msg_id: u64,
-    committed_started_at: &str,
-    committed_turn_start_offset: Option<u64>,
-) -> usize {
-    let applier = shared_reaction_applier(shared.clone());
-    drain_on_terminal_commit_with_applier_and_offset(
-        provider,
-        tmux_session_name,
-        channel_id,
-        now_ms(),
-        committed_user_msg_id,
-        committed_started_at,
-        committed_turn_start_offset,
-        &applier,
-    )
-    .await
-}
-
-pub(super) async fn drain_on_terminal_commit_with_offsets(
-    shared: &Arc<SharedData>,
-    provider: &str,
-    tmux_session_name: &str,
-    channel_id: u64,
-    committed_user_msg_id: u64,
-    committed_started_at: &str,
-    committed_turn_start_offset: Option<u64>,
-    committed_terminal_evidence_offset: Option<u64>,
-) -> usize {
-    let applier = shared_reaction_applier(shared.clone());
-    drain_on_terminal_commit_with_applier_and_offsets(
-        provider,
-        tmux_session_name,
-        channel_id,
-        now_ms(),
-        committed_user_msg_id,
-        committed_started_at,
-        committed_turn_start_offset,
-        TerminalEvidenceOffsetProof::Recorded(committed_terminal_evidence_offset),
-        &applier,
-    )
-    .await
-}
-
-pub(super) async fn drain_on_terminal_commit_with_applier(
-    provider: &str,
-    tmux_session_name: &str,
-    channel_id: u64,
-    now_ms: u64,
-    committed_user_msg_id: u64,
-    committed_started_at: &str,
-    applier: &ReactionApplierFn,
-) -> usize {
-    drain_on_terminal_commit_with_applier_and_offset(
-        provider,
-        tmux_session_name,
-        channel_id,
-        now_ms,
-        committed_user_msg_id,
-        committed_started_at,
-        None,
-        applier,
-    )
-    .await
-}
-
-pub(super) async fn drain_on_terminal_commit_with_applier_and_offset(
-    provider: &str,
-    tmux_session_name: &str,
-    channel_id: u64,
-    now_ms: u64,
-    committed_user_msg_id: u64,
-    committed_started_at: &str,
-    committed_turn_start_offset: Option<u64>,
-    applier: &ReactionApplierFn,
-) -> usize {
-    drain_on_terminal_commit_with_applier_and_offsets(
-        provider,
-        tmux_session_name,
-        channel_id,
-        now_ms,
-        committed_user_msg_id,
-        committed_started_at,
-        committed_turn_start_offset,
-        TerminalEvidenceOffsetProof::Legacy,
-        applier,
-    )
-    .await
-}
-
-async fn drain_on_terminal_commit_with_applier_and_offsets(
-    provider: &str,
-    tmux_session_name: &str,
-    channel_id: u64,
-    now_ms: u64,
-    committed_user_msg_id: u64,
-    committed_started_at: &str,
-    committed_turn_start_offset: Option<u64>,
-    terminal_evidence_offset: TerminalEvidenceOffsetProof,
-    applier: &ReactionApplierFn,
-) -> usize {
-    let mut drained = 0usize;
-    for marker in load_for_channel(provider, channel_id) {
-        if marker.tmux_session_name != tmux_session_name {
-            continue; // I4: a different session's marker is never this commit's
-        }
-        // Mutual exclusion vs the sweep (verify r1 fix #2): claim, then decide
-        // on a FRESH re-read — the sweep may have resolved it since load.
-        let Some(_claim) = try_claim_marker(&marker) else {
-            continue; // the sweep owns this marker right now
-        };
-        let Some(mut marker) = reload(&marker) else {
-            continue; // resolved while unclaimed
-        };
-        if !terminal_commit_covers_marker_with_offsets(
-            now_ms,
-            &marker,
-            committed_user_msg_id,
-            committed_started_at,
-            committed_turn_start_offset,
-            terminal_evidence_offset,
-        ) {
-            continue;
-        }
-        match applier(&marker, ReactionOp::Complete).await {
-            ReactionDelivery::Delivered => {
-                delete(&marker);
-                drained += 1;
-                tracing::info!(
-                    provider = %marker.provider,
-                    channel_id = marker.channel_id,
-                    tmux_session_name = %marker.tmux_session_name,
-                    anchor_message_id = marker.anchor_message_id,
-                    origin = ?marker.origin,
-                    "tui_direct_abort_marker: anchor covered by the pinned turn's terminal commit; ⏳ → ✅ delivered and marker drained (#3296/#3303)"
-                );
-            }
-            ReactionDelivery::FailedPermanent => {
-                // Permanently gone anchor (404/403/410): no reaction can EVER
-                // land; terminate, don't retry (verify r1 fix #3). WARN here only.
-                delete(&marker);
-                tracing::warn!(
-                    provider = %marker.provider,
-                    channel_id = marker.channel_id,
-                    anchor_message_id = marker.anchor_message_id,
-                    "tui_direct_abort_marker: anchor permanently gone (404/403/410); covered marker terminated without ✅ (#3296)"
-                );
-            }
-            ReactionDelivery::Failed => {
-                // I6 fail-open: the anchor IS covered — stamp it so the sweep
-                // retries the ✅ (and can never degrade it to ⚠).
-                marker.covered_at_ms = Some(now_ms);
-                if let Err(error) = record(&marker) {
-                    // verify r1 fix #4: surface loudly — a swallowed stamp
-                    // failure would let the sweep ⚠ a COVERED anchor after the
-                    // TTL; un-stamped, the next covering drain retries the ✅.
-                    tracing::error!(
-                        provider = %marker.provider,
-                        channel_id = marker.channel_id,
-                        anchor_message_id = marker.anchor_message_id,
-                        error = %error,
-                        "tui_direct_abort_marker: failed to persist covered_at stamp after a ✅ delivery failure; next covering commit retries (#3296)"
-                    );
-                }
-            }
-        }
-    }
-    drained
-}
-
-/// #3350 issue-1 (lease-gated row-absent commit): a VISIBLE `⏳ → ✅` was just
-/// delivered on `anchor_message_id`, but the committed row was already gone,
-/// so the watcher chokepoint has no row identity to tombstone. Source it from
-/// the marker instead: every DeferredClaim marker pinned to ITS OWN anchor
-/// (#3303 SC1) gets its pinned identity recorded as a commit tombstone FIRST
-/// (write-before-discard — a sweep claiming the marker mid-pass still 대조s
-/// `✅`, never `⚠`), then is discarded under the claim mutex. Reaction-free by
-/// design (#3350 I1): the `✅` is already on the message, so the marker's
-/// whole job — bounding the `⏳` — is done. Abort-kind markers (foreign pin)
-/// are left untouched: this commit proves nothing about the foreign prior
-/// turn, so their #3296 convergence (drain / sweep / hard cap) is preserved.
-pub(super) fn resolve_own_claim_markers_for_visibly_completed_anchor(
-    provider: &str,
-    tmux_session_name: &str,
-    channel_id: u64,
-    anchor_message_id: u64,
-) -> usize {
-    let mut resolved = 0usize;
-    for marker in load_for_channel(provider, channel_id) {
-        if marker.tmux_session_name != tmux_session_name
-            || marker.anchor_message_id != anchor_message_id
-            || marker.origin != MarkerOrigin::DeferredClaim
-            || marker.foreign_user_msg_id != Some(anchor_message_id)
-        {
-            continue; // SC1: own-pin DeferredClaim markers of THIS anchor only
-        }
-        let Some(own_started_at) = marker.foreign_started_at.clone() else {
-            continue; // identity-absent marker — the sweep bound alone resolves it
-        };
-        record_commit_tombstone_with_offset(
-            provider,
-            tmux_session_name,
-            channel_id,
-            anchor_message_id,
-            &own_started_at,
-            marker.foreign_turn_start_offset,
-        );
-        let Some(_claim) = try_claim_marker(&marker) else {
-            continue; // a reconciler owns it; the durable tombstone keeps its verdict ✅
-        };
-        let Some(marker) = reload(&marker) else {
-            continue; // resolved while unclaimed
-        };
-        delete(&marker);
-        resolved += 1;
-        tracing::info!(
-            provider = %marker.provider,
-            channel_id = marker.channel_id,
-            tmux_session_name = %marker.tmux_session_name,
-            anchor_message_id = marker.anchor_message_id,
-            "tui_direct_abort_marker: own-pin marker discarded after a visible lease-gated ✅ (row-absent commit); tombstone recorded first so a racing sweep still lands ✅ (#3350)"
-        );
-    }
-    resolved
-}
-
-/// Placeholder-sweeper pass: retry `✅` for covered markers; apply the TTL'd
-/// `⏳ → ⚠` fallback for anchors no commit ever covered (held while a live
-/// inflight may still cover them, up to the hard cap). Returns resolved.
-pub(super) async fn sweep_expired(
-    shared: &Arc<SharedData>,
-    provider: &super::ProviderKind,
-) -> usize {
-    let http_available = shared.serenity_http_or_token_fallback().is_some();
-    let applier = shared_reaction_applier(shared.clone());
-    // ONE inflight read feeds both kind predicates (#3303): `defers` is the
-    // Abort-kind hold; `is_pinned_turn` is the DeferredClaim-kind hold (the
-    // live row IS the marker's pinned own turn — user_msg_id / started_at /
-    // tmux session name all match).
-    let live_inflight = |marker: &AbortedAnchorMarker| -> LiveInflightProbe {
-        match super::inflight::load_inflight_state(provider, marker.channel_id) {
-            None => LiveInflightProbe {
-                defers: false,
-                is_pinned_turn: false,
-            },
-            Some(state) => LiveInflightProbe {
-                defers: inflight_defers_sweep(
-                    marker,
-                    state.tmux_session_name.as_deref(),
-                    state.user_msg_id,
-                    &state.started_at,
-                ),
-                is_pinned_turn: deferred_claim_live_inflight_is_pinned(
-                    marker,
-                    state.tmux_session_name.as_deref(),
-                    state.user_msg_id,
-                    &state.started_at,
-                    state.turn_start_offset,
-                ),
-            },
-        }
-    };
-    sweep_expired_with_applier(
-        provider.as_str(),
-        now_ms(),
-        http_available,
-        &live_inflight,
-        &applier,
-    )
-    .await
-}
-
-/// Generic over the probe return (`Into<LiveInflightProbe>`) so the existing
-/// Abort-kind test closures keep their bare-bool form (`From<bool>` maps to a
-/// never-pinned probe) while the production sweep supplies the full #3303
-/// probe.
-pub(super) async fn sweep_expired_with_applier<P: Into<LiveInflightProbe>>(
-    provider: &str,
-    now_ms: u64,
-    http_available: bool,
-    live_inflight_for_session: &(dyn Fn(&AbortedAnchorMarker) -> P + Send + Sync),
-    applier: &ReactionApplierFn,
-) -> usize {
-    let mut resolved = 0usize;
-    for marker in load_all() {
-        if !marker.provider.eq_ignore_ascii_case(provider) {
-            continue;
-        }
-        if marker.anchor_message_id == 0 {
-            delete(&marker); // I5: corrupt record — nothing could ever target it
-            continue;
-        }
-        // Mutual exclusion vs the watcher drain (verify r1 fix #2): claim,
-        // then decide on a FRESH re-read — a terminal commit racing this pass
-        // may have covered or drained the marker since load.
-        let Some(_claim) = try_claim_marker(&marker) else {
-            continue; // the drain owns this marker right now
-        };
-        let Some(mut marker) = reload(&marker) else {
-            continue; // drained while unclaimed
-        };
-        // codex r2 ordering: live-inflight read FIRST, tombstone 대조 SECOND.
-        // The chokepoint writes the tombstone BEFORE clearing the row, so "no
-        // live row" here guarantees a commit-caused clear already made its
-        // tombstone visible to the 대조 — a sweep pass claiming the marker
-        // mid-commit can no longer beat the correct ✅ with a ⚠ (finding 1;
-        // the flock only serializes, it does not order the verdicts).
-        let probe: LiveInflightProbe = live_inflight_for_session(&marker).into();
-        if marker.covered_at_ms.is_none()
-            && let Some(committed_at_ms) = post_abort_commit_tombstone(&marker)
-        {
-            marker.covered_at_ms = Some(committed_at_ms);
-            if let Err(error) = record(&marker) {
-                // Loud (verify-r1 fix #4): un-persisted, the stamp re-derives
-                // from the tombstone next pass — but a Discord outage outliving
-                // the tombstone retention would then ⚠ a covered anchor.
-                tracing::error!(
-                    provider = %marker.provider,
-                    channel_id = marker.channel_id,
-                    anchor_message_id = marker.anchor_message_id,
-                    error = %error,
-                    "tui_direct_abort_marker: failed to persist tombstone-covered stamp; next sweep pass re-derives (#3296 r2)"
-                );
-            }
-        }
-        // #3303: ONLY the disposition branches on the marker kind. The Abort
-        // disposition (incl. its 6×TTL hard cap) is byte-for-byte untouched;
-        // the DeferredClaim disposition swaps the live-inflight hold for the
-        // pinned-own-turn hold (uncapped) and drops the name-only hold.
-        let disposition = match marker.origin {
-            MarkerOrigin::Abort => decide_marker_disposition(
-                now_ms,
-                &marker,
-                probe.defers,
-                ABORT_MARKER_TTL,
-                http_available,
-            ),
-            MarkerOrigin::DeferredClaim => {
-                deferred_claim::decide_deferred_claim_marker_disposition(
-                    now_ms,
-                    &marker,
-                    probe,
-                    ABORT_MARKER_TTL,
-                    http_available,
-                )
-            }
-        };
-        let op = match disposition {
-            MarkerDisposition::KeepWaiting | MarkerDisposition::LeftIntactHttpUnavailable => {
-                continue;
-            }
-            MarkerDisposition::DeliverCompletion => ReactionOp::Complete,
-            MarkerDisposition::DeliverFailureWarn => ReactionOp::FailureWarn,
-        };
-        match applier(&marker, op).await {
-            ReactionDelivery::Delivered => {
-                delete(&marker);
-                resolved += 1;
-                tracing::info!(
-                    provider = %marker.provider,
-                    channel_id = marker.channel_id,
-                    tmux_session_name = %marker.tmux_session_name,
-                    anchor_message_id = marker.anchor_message_id,
-                    op = ?op,
-                    origin = ?marker.origin,
-                    "tui_direct_abort_marker: sweep resolved marked anchor (#3296/#3303)"
-                );
-            }
-            ReactionDelivery::FailedPermanent => {
-                // Permanently gone anchor (404/403/410): terminate the marker
-                // instead of retrying every pass forever (verify r1 fix #3).
-                delete(&marker);
-                tracing::warn!(
-                    provider = %marker.provider,
-                    channel_id = marker.channel_id,
-                    anchor_message_id = marker.anchor_message_id,
-                    op = ?op,
-                    "tui_direct_abort_marker: anchor permanently gone (404/403/410); marker terminated by sweep (#3296)"
-                );
-            }
-            // I6: keep the marker for the next pass (delivery failed late).
-            ReactionDelivery::Failed => {}
-        }
-    }
-    // GC AFTER the marker loop: the first pass after long downtime must 대조
-    // against evidence that aged past the retention cap before deleting it.
-    gc_expired_commit_tombstones(now_ms);
-    resolved
 }
 
 #[cfg(test)]
